perf/ethp: Use '--minr 3' option for better performance
With higher '--minr' value, DAMON becomes more accurate.  In case of
ETHP, this will make the MADV_HUGEPAGE hint to be applied to the memory
region _really_ accessed.  In other words, the pages for the region are
already faulted with the regular pages.  Therefore, number of thp fault
alloc will be decreased.  The collapsing of khugepaged is only one
source of the benefit now.  However, the khugepaged scanning is not so
aggressive (iteratively scans 4096 pages and sleeps 10 seconds).  This
might be enough for long live processes, but the benchmark program of
ours would not get sufficient performance benefit.  For the reason, this
commit makes ETHP run to set '--minr' value as 3.

Signed-off-by: SeongJae Park <sjpark@amazon.de>