perf: Use record format
On kernels doesn's supporting the record feature of DAMON, DAMO stores
the monitoring results in 'perf script' output format, so the size of
the results files become quite big.  This commit avoids the space
overhead by making the test convert the final result file's format to
that of the record feature.

Signed-off-by: SeongJae Park <sj@kernel.org>